requirements19.txt
bottlenose==0.6.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.4
django-admin-bootstrapped==2.4.0
django-admin-tools==0.7.2
django-appconf==0.6
django-autocomplete-light==3.1.4
django-bootstrap-form==3.1
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-follow==0.6.1
django-geoposition==0.2.2
django-grappelli==2.8.1
django-guardian==1.2.5
django-haystack==2.4.1
django-jquery==1.9.1
django-markdown-deux==1.0.5
django-mathfilters==0.3.0
django-model-utils==2.2
django-profiles==0.2
django-registration-redux==1.4
django-sortedm2m==0.9.4
django-tagging==0.4.1
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
djrill==1.4.0
easy-thumbnails==2.2
gunicorn==19.4.3
imagestore==2.9.1
lxml==3.4.4
markdown2==2.3.0
Pillow==3.0.0
pinax-theme-bootstrap==5.7.1
psycopg2==2.6.1
pygeoip==0.3.2
pyreadline==2.0
pytz==2014.10
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.3
sqlparse==0.1.19
static3==0.6.1
swapper==0.3.0
whitenoise==3.0
